+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changelog
 * Record: 增加forUpdate,forShare,lock方法用于事务中的锁
 * Db: insertBatch增加extra参数,可用于在生成的语句后面增加"ON DUPLICATE KEY UPDATE"字句
 * Http: 网络请求失败时,抛出的异常code加2000(如404则异常code为2404),避免被error服务认为是当前页面404
+* App: 自定义控制器格式中的命名空间,要求大写形式
 
 ## 0.9.22 (2016-11-22)